// src/types/apiMessages.ts
export enum ApiMessageEnum {
  REGISTER_PERSONAL_SUCCESS = 'MESSAGE_REGISTER_PERSONAL_SUCCESS',
  REGISTER_COMPANY_SUCCESS = 'MESSAGE_REGISTER_COMPANY_SUCCESS',
  LOGIN_PERSONAL_SUCCESS = 'MESSAGE_LOGIN_PERSONAL_SUCCESS',
  LOGIN_COMPANY_SUCCESS = 'MESSAGE_LOGIN_COMPANY_SUCCESS',
  LOGIN_COMPANY_FAIL = 'MESSAGE_LOGIN_COMPANY_FAIL',
  REFRESH_TOKENS_SUCCESS = 'MESSAGE_REFRESH_TOKENS_SUCCESS',
  IS_DRAFT_USER = 'MESSAGE_IS_DRAFT_USER',
  VERIFY_EMAIL_SUCCESS = 'MESSAGE_VERIFY_EMAIL_SUCCESS',
  LOGOUT_SUCCESS = 'LOGOUT_SUCCESS',
  CREATE_ADMIN_SUCCESS = 'MESSAGE_CREATE_ADMIN_SUCCESS',
  LOGIN_ADMIN_SUCCESS = 'MESSAGE_LOGIN_ADMIN_SUCCESS',
  ADMIN_CREATE_POST_SUCCESS = 'MESSAGE_ADMIN_CREATE_POST_SUCCESS',
  ADMIN_UPDATE_POST_SUCCESS = 'MESSAGE_ADMIN_UPDATE_POST_SUCCESS',
  ADMIN_DELETE_POST_SUCCESS = 'MESSAGE_ADMIN_DELETE_POST_SUCCESS',
  FIND_POST_SUCCESS = 'MESSAGE_FIND_POST_SUCCESS',
  UPLOAD_POST_THUMBNAIL_SUCCESS = 'MESSAGE_UPLOAD_POST_THUMBNAIL_SUCCESS',
  FETCH_SUCCESS = 'MESSAGE_FETCH_SUCCESS',
  UPLOAD_NEWS_BRAND_THUMBNAIL_SUCCESS = 'MESSAGE_UPLOAD_NEWS_BRAND_THUMBNAIL_SUCCESS',
  CREATED_SUCCESS = 'MESSAGE_CREATED_SUCCESS',
  UPLOADED_SUCCESS = 'MESSAGE_UPLOADED_SUCCESS',
  UPDATED_SUCCESS = 'MESSAGE_UPDATED_SUCCESS',
  DELETED_SUCCESS = 'MESSAGE_DELETED_SUCCESS',
  CREATE_GOOGLE_LOGIN_URL_SUCCESS = 'MESSAGE_CREATE_GOOGLE_LOGIN_URL_SUCCESS',
  LOGIN_GOOGLE_SUCCESS = 'MESSAGE_LOGIN_GOOGLE_SUCCESS',
  FOLLOW_SUCCESS = 'MESSAGE_FOLLOW_SUCCESS',
  CORRECT_PASSWORD = 'MESSAGE_CORRECT_PASSWORD',
}